requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
yt-dlp>=2024.8.6
google-genai>=1.0.0
python-multipart==0.0.9
Pillow==10.4.0
